+++ packages/algolia-mesh/mesh/algoliaFacetsToAggregations.ts
@@ -1,6 +1,7 @@
 import type {
   Aggregation,
   AggregationOption,
+  AlgoliasearchResponse,
   CategoryResult,
   MeshContext,
 } from '@graphcommerce/graphql-mesh'
@@ -23,11 +24,7 @@ function categoryMapping(
       const count = category?.id ? facetList[category?.id] : 0
       return { label: category?.name, value: category?.uid ?? '', count }
     })
-    .filter((category) => {
-      if (category.count > 0) {
-        return category
-      }
-    })
+    .filter((category) => category.count > 0)
 }
 
 function compare(a, b) {
@@ -122,10 +119,11 @@ function assertAlgoliaFacets(facets: any): facets is AlgoliaFacets {
  * TODO: Make sure the aggregations are sorted correctly: https://magento-247-git-canary-graphcommerce.vercel.app/men/photography, through position
  */
 export function algoliaFacetsToAggregations(
-  algoliaFacets: any,
+  algoliaFacets: AlgoliasearchResponse['facets'],
   attributes: AttributeList,
   storeConfig: GetStoreConfigReturn,
   categoryList?: null | CategoryResult,
+  groupId?: number,
 ): Aggregation[] {
   if (!storeConfig?.default_display_currency_code) throw new Error('Currency is required')
   const aggregations: Aggregation[] = []
@@ -152,6 +150,15 @@ export function algoliaFacetsToAggregations(
         position,
       })
     } else if (facetIndex.startsWith('price')) {
+      if (!groupId && facetIndex !== `price.${storeConfig.default_display_currency_code}.default`) {
+        return
+      }
+      if (
+        groupId &&
+        facetIndex !== `price.${storeConfig.default_display_currency_code}.group_${groupId}`
+      ) {
+        return
+      }
       aggregations.push({
         label,
         attribute_code,
@@ -176,8 +183,14 @@ export function algoliaFacetsToAggregations(
   return aggregations
 }
 
-export function getCategoryList(context: MeshContext) {
-  return context.m2.Query.categories({
+let categoryListCache: CategoryResult | null = null
+
+export async function getCategoryList(context: MeshContext) {
+  if (categoryListCache) return categoryListCache
+
+  // context.cache.get('algolia_getCategoryList')
+
+  categoryListCache = await context.m2.Query.categories({
     args: { filters: {} },
     selectionSet: /* GraphQL */ `
       {
@@ -191,4 +204,6 @@ export function getCategoryList(context: MeshContext) {
     `,
     context,
   })
+
+  return categoryListCache!
 }