+++ packages/algolia-mesh/mesh/getAlgoliaSettings.ts
@@ -1,20 +1,21 @@
-import { AlgoliaindexSettings, MeshContext } from '@graphcommerce/graphql-mesh'
+import { AlgoliasettingsResponse, MeshContext } from '@graphcommerce/graphql-mesh'
 import { getIndexName } from './getIndexName'
 
-export async function getAlgoliaSettings(context: MeshContext): Promise<AlgoliaindexSettings> {
-  const settings = await context.algolia.Query.algolia_getSettings({
-    args: { indexName: getIndexName(context) },
-    selectionSet: /* GraphQL */ `
-      {
-        replicas
-      }
-    `,
-    context,
-  })
+let settingsCache: AlgoliasettingsResponse | null = null
 
-  if (!settings) {
-    throw new Error('No settings found')
+export async function getAlgoliaSettings(context: MeshContext): Promise<AlgoliasettingsResponse> {
+  if (!settingsCache) {
+    settingsCache = await context.algolia.Query.algolia_getSettings({
+      args: { indexName: getIndexName(context) },
+      selectionSet: /* GraphQL */ `
+        {
+          replicas
+        }
+      `,
+      context,
+    })
   }
 
-  return settings
+  if (!settingsCache) throw Error('No settings found')
+  return settingsCache
 }