+++ packages/algolia-mesh/mesh/getAttributeList.ts
@@ -4,16 +4,25 @@ import { filterNonNullableKeys } from '@graphcommerce/next-ui/RenderType/filterN
 
 export type AttributeList = { label: string; code: string }[]
 
+let attributeListCache: AttributeList | null = null
+
 export async function getAttributeList(context: MeshContext): Promise<AttributeList> {
-  if (import.meta.graphCommerce.magentoVersion >= 247 && 'attributesList' in context.m2.Query && typeof context.m2.Query.attributesList === 'function') {
-    const items = await context.m2.Query.attributesList({
+  if (attributeListCache) return attributeListCache
+
+  if (
+    import.meta.graphCommerce.magentoVersion >= 247 &&
+    'attributesList' in context.m2.Query &&
+    typeof context.m2.Query.attributesList === 'function'
+  ) {
+    const items = (await context.m2.Query.attributesList({
       args: {
         entityType: 'CATALOG_PRODUCT',
         filters: {},
       },
       selectionSet: `{ items{ code label } }`,
       context,
-    }).then((res) => res?.items) as { label?: string; code: string }[]
+    }).then((res) => res?.items)) as { label?: string; code: string }[]
+    attributeListCache = filterNonNullableKeys(items, ['label'])
 
     return filterNonNullableKeys(items, ['label'])
   }