+++ packages/algolia-mesh/mesh/productFilterInputToAlgoliafacetFiltersInput.ts
@@ -1,6 +1,5 @@
 import {
   ProductAttributeFilterInput,
-  AlgoliafacetFilters_Input,
   AlgolianumericFilters_Input,
 } from '@graphcommerce/graphql-mesh'
 import {
@@ -20,22 +19,23 @@ import { nonNullable } from './utils'
 export function productFilterInputToAlgoliaFacetFiltersInput(
   filters?: InputMaybe<ProductAttributeFilterInput>,
 ) {
-  const filterArray: AlgoliafacetFilters_Input[] = []
+  const filterArray: (string | string[])[] = []
   if (!filters) {
     return []
   }
 
   Object.entries(filters).forEach(([key, value]) => {
     if (isFilterTypeEqual(value)) {
-      const valueArray = ((value?.in ? value?.in : [value?.eq]) ?? []).filter(nonNullable)
+      if (value.in) {
+        const values = value.in.filter(nonNullable)
+        if (key === 'category_uid') filterArray.push(values.map((v) => `categoryIds:${atob(v)}`))
+        else filterArray.push(values.map((v) => `${key}:${v}`))
+      }
 
-      valueArray.forEach((v) => {
-        if (key === 'category_uid') {
-          filterArray.push({ facetFilters_Input: { String: `categoryIds:${atob(v)}` } })
-        } else {
-          filterArray.push({ facetFilters_Input: { String: `${key}:${v}` } })
-        }
-      })
+      if (value.eq) {
+        if (key === 'category_uid') filterArray.push(`categoryIds:${atob(value.eq)}`)
+        else filterArray.push(`${key}:${value.eq}`)
+      }
     }
 
     if (isFilterTypeMatch(value)) {
@@ -51,49 +51,31 @@ export function productFilterInputToAlgoliaFacetFiltersInput(
  *
  * https://www.algolia.com/doc/api-reference/api-parameters/numericFilters/#examples
  */
-export function productFilterInputToAlgoliaNumericFiltersInput(
-  storeConfig: GetStoreConfigReturn,
+export async function productFilterInputToAlgoliaNumericFiltersInput(
+  storeConfig: Promise<GetStoreConfigReturn>,
   filters?: InputMaybe<ProductAttributeFilterInput>,
 ) {
-  const filterArray: AlgolianumericFilters_Input[] = []
-  if (!storeConfig) {
-    throw Error('StoreConfig is missing')
-  }
-  if (!filters) {
-    return []
-  }
+  if (!filters) return []
 
-  Object.entries(filters).forEach(([key, value]) => {
+  const filterArray: AlgolianumericFilters_Input[] = []
+
+  for (const [key, value] of Object.entries(filters)) {
     if (isFilterTypeRange(value)) {
       if (key === 'price') {
+        // eslint-disable-next-line no-await-in-loop
+        const currencyCode = (await storeConfig)?.default_display_currency_code
         filterArray.push(
-          {
-            numericFilters_Input: {
-              String: `${key}.${storeConfig.default_display_currency_code}.default >= ${value.from}`,
-            },
-          },
-          {
-            numericFilters_Input: {
-              String: `${key}.${storeConfig.default_display_currency_code}.default <= ${value.to}`,
-            },
-          },
+          { numericFilters_Input: { String: `${key}.${currencyCode}.default >= ${value.from}` } },
+          { numericFilters_Input: { String: `${key}.${currencyCode}.default <= ${value.to}` } },
         )
       } else {
         filterArray.push(
-          {
-            numericFilters_Input: {
-              String: `${key} >= ${value.from}`,
-            },
-          },
-          {
-            numericFilters_Input: {
-              String: `${key} <= ${value.to}`,
-            },
-          },
+          { numericFilters_Input: { String: `${key} >= ${value.from}` } },
+          { numericFilters_Input: { String: `${key} <= ${value.to}` } },
         )
       }
     }
-  })
+  }
 
   return filterArray
 }