+++ packages/google-datalayer/plugins/GoogleDatalayerPaymentMethodContextProvider.tsx
@@ -1,7 +1,7 @@
 import type { PaymentMethodContextProviderProps } from '@graphcommerce/magento-cart-payment-method'
 import type { PluginConfig, PluginProps } from '@graphcommerce/next-config'
-import { sendEvent } from '../api/sendEvent'
-import { orderToPurchase } from '../mapping/orderToPurchase/orderToPurchase'
+import { useSendEvent } from '../api/sendEvent'
+import { cartToPurchase } from '../mapping/cartToPurchase/cartToPurchase'
 
 export const config: PluginConfig = {
   module: '@graphcommerce/magento-cart-payment-method',
@@ -12,11 +12,13 @@ export function PaymentMethodContextProvider(
   props: PluginProps<PaymentMethodContextProviderProps>,
 ) {
   const { Prev, onSuccess, ...rest } = props
+
+  const sendEvent = useSendEvent()
   return (
     <Prev
       {...rest}
       onSuccess={(orderNumber, cart) => {
-        sendEvent('purchase', orderToPurchase(orderNumber, cart))
+        sendEvent('purchase', cartToPurchase(orderNumber, cart))
         return onSuccess?.(orderNumber, cart)
       }}
     />