+++ packagesDev/next-config/__tests__/config/utils/mergeEnvIntoConfig.ts
@@ -84,6 +84,11 @@ it('converts an env schema to a config schema', () => {
     magentoEndpoint: 'https://example.com',
     previewSecret: 'secret',
     magentoVersion: 247,
+    algolia: {
+      applicationId: 'a',
+      indexNamePrefix: 'a',
+      searchOnlyApiKey: 'a',
+    },
   }
 
   const environmentVariables = {